requirements.txt
python-nmap
scapy
cryptography
requests
jinja2
psutil
pywin32  # Solo si usas Windows
PyYAML
